requirements.txt
streamlit
ollama
PyPDF2
python-docx
